hisi_hal::dma

Struct Sdma0

pub struct Sdma0;

Marker type for the secure DMA controller (logical channels 8-11).

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.